+++ src/process/validation/__tests__/multiple.test.ts
@@ -1,10 +1,13 @@
 import { expect } from 'chai';
+import { ValidationContext } from 'types';
 import { validationRules } from '..';
 import { rules, serverRules } from '../rules';
+import { shouldValidate as shouldValidateRegexPattern } from '../rules/validateRegexPattern';
+import { shouldValidate as shouldValidateRequired } from '../rules/validateRequired';
 
 const allRules = [...rules, ...serverRules];
 
-const component = {
+const textFieldComponent = {
     type: 'textfield',
     key: 'multiple_textfield',
     label: 'Multiple Textfield',
@@ -17,9 +20,36 @@ const component = {
         pattern: '^[0-9]+$',
     }
 };
+const selectComponent = {
+    type: 'select',
+    key: 'multiple_select',
+    label: 'Multiple Select',
+    widget: 'choicesjs',
+    input: true,
+    multiple: true,
+    data: {
+        values: [
+            {
+                label: 'A',
+                value: 'a',
+            },
+            {
+                label: 'B',
+                value: 'b',
+            },
+            {
+                label: 'C',
+                value: 'c',
+            },
+        ],
+    },
+    validate: {
+        required: true,
+    }
+};
 
-const context = {
-    component,
+const context: ValidationContext = {
+    component: textFieldComponent,
     value: [],
     path: 'multiple_textfield',
     data: {multiple_textfield: []},
@@ -27,14 +57,38 @@ const context = {
     scope: {errors: []},
 };
 
+const contextWithSelectComponent: ValidationContext = {
+    component: selectComponent,
+    value: [],
+    path: 'multiple_select',
+    data: {multiple_select: []},
+    row: {multiple_select: []},
+    scope: {errors: []},
+};
+
 it('Validating required rule will work for multiple values component with no rows', async () => {
-    const fullValueRules = allRules.filter((rule) => rule.fullValue);
-    const rulesToValidate = validationRules(context, fullValueRules, undefined);
-    expect(rulesToValidate).to.not.have.length(0);
+    // TextField
+    const shouldValidateRequiredTextField = shouldValidateRequired(context);
+    expect(shouldValidateRequiredTextField).to.be.equal(true);
+    // Select
+    const shouldValidateRequiredSelect = shouldValidateRequired(contextWithSelectComponent);
+    expect(shouldValidateRequiredSelect).to.be.equal(true);
 });
 
-it('Validati olther rules will skip for multiple values component with no rows', async () => {
-    const otherRules = allRules.filter((rule) => !rule.fullValue);  
-    const rulesToValidate = validationRules(context, otherRules, undefined);
-    expect(rulesToValidate).to.have.length(0);
+it('Validate RegexPattern rule won\'t execute for multiple values component with no rows', async () => {
+    const shouldValidateRegexRule = shouldValidateRegexPattern(context);
+    expect(shouldValidateRegexRule).to.be.equal(false);
+
+    const contextWithValues = {
+        ...context,
+        value: 'a',
+        data: {
+            multiple_textfield: ['a'],
+        },
+        row: {
+            multiple_textfield: ['a'],
+        },
+    };
+    const shouldValidateRegexRuleWithValues = shouldValidateRegexPattern(contextWithValues);
+    expect(shouldValidateRegexRuleWithValues).to.be.equal(true);
 });