+++ src/process/validation/index.ts
@@ -7,7 +7,6 @@ import { getComponentAbsolutePath, getComponentPath } from "utils/formUtil";
 import { getErrorMessage } from "utils/error";
 import { FieldError } from "error";
 import { ConditionallyHidden, isConditionallyHidden, isCustomConditionallyHidden, isSimpleConditionallyHidden } from "processes/conditions";
-import { isParentHidden } from 'utils/isParentHidden';
 
 // Cleans up validation errors to remove unnessesary parts
 // and make them transferable to ivm.
@@ -48,14 +47,6 @@ export function validationRules(
     }
     const validationRules: ValidationRuleInfo[] = [];
     return rules.reduce((acc, rule: ValidationRuleInfo) => {
-        if (context.component.multiple &&
-            Array.isArray(context.value) &&
-            context.value?.length === 0 &&
-            !rule.fullValue
-        ) {
-            return acc;
-        }
-
         if (rule.shouldProcess && rule.shouldProcess(context)) {
             acc.push(rule);
         }
@@ -98,10 +89,10 @@ export const _shouldSkipValidation = (context: ValidationContext, isConditionall
 
     if (
         (scope as ConditionsScope)?.conditionals &&
-        find((scope as ConditionsScope).conditionals, {
+        (find((scope as ConditionsScope).conditionals, {
             path: getComponentPath(component, path),
             conditionallyHidden: true
-        })
+        }) || component.ephemeralState?.conditionallyHidden === true)
     ) {
         return true;
     }