+++ src/utils/conditions.ts
@@ -1,7 +1,7 @@
 import { ConditionsContext, JSONConditional, LegacyConditional, SimpleConditional } from "types";
 import { EvaluatorFn, evaluate, JSONLogicEvaluator } from 'modules/jsonlogic';
-import { getComponent, getComponentActualValue } from "./formUtil";
-import { has, isObject, map, every, some, find, filter } from 'lodash';
+import { flattenComponents, getComponent, getComponentActualValue } from "./formUtil";
+import { has, isObject, map, every, some, find, filter, isBoolean, split } from 'lodash';
 import ConditionOperators from './operators';
 import _ from 'lodash';
 
@@ -18,7 +18,7 @@ export const isSimpleConditional = (conditional: any): conditional is SimpleCond
 }
 
 export function conditionallyHidden(context: ConditionsContext) {
-    const { scope, component, path } = context;
+    const { scope, path } = context;
     if (scope.conditionals && path) {
         const hidden = find(scope.conditionals, (conditional) => {
             return conditional.path === path;
@@ -135,13 +135,29 @@ function getConditionalPathsRecursive(conditionPaths: any, data: any): any {
 }
 
 /**
+ * Checks if condition can potentially have a value path instead of component path.
+ * @param condition
+ * @returns {boolean}
+ */
+function isConditionPotentiallyBasedOnValuePath(condition: any = {}) {
+    let comparedValue;
+    try {
+        comparedValue = JSON.parse(condition.value);
+    }
+    catch(e) {
+        comparedValue = condition.value;
+    }
+    return isBoolean(comparedValue) && (condition.component || '').split('.').length > 1 && condition.operator === 'isEqual';
+}
+
+/**
  * Checks the simple conditionals.
  * @param conditional
  * @param context
  * @returns
  */
 export function checkSimpleConditional(conditional: SimpleConditional, context: ConditionsContext): boolean | null {
-    const { component, data, row, instance, form, components = [] } = context;
+    const { component, data, row, instance, form } = context;
     if (!conditional || !isSimpleConditional(conditional)) {
         return null;
     }
@@ -173,12 +189,39 @@ export function checkSimpleConditional(conditional: SimpleConditional, context: 
         }
 
         const conditionsResult = filter(map(conditions, (cond) => {
-            const { value: comparedValue, operator, component: conditionComponentPath } = cond;
+            let { value: comparedValue, operator, component: conditionComponentPath } = cond;
             if (!conditionComponentPath) {
                 // Ignore conditions if there is no component path.
                 return null;
             }
-            const conditionComponent = getComponent(form?.components || components, conditionComponentPath, true);
+
+            const formComponents = form?.components || [];
+            let conditionComponent = getComponent(formComponents, conditionComponentPath, true);
+            // If condition componenet is not found, check if conditionComponentPath is value path.
+            // Need to handle condtions like:
+            //   {
+            //     "component": "selectBoxes.a",
+            //     "operator": "isEqual",
+            //     "value": "true"
+            //   }
+            if (!conditionComponent && isConditionPotentiallyBasedOnValuePath(cond) && formComponents.length) {
+                const flattenedComponents = flattenComponents(formComponents, true);
+                const pathParts = split(conditionComponentPath, '.');
+                const valuePathParts = [];
+
+                while (!conditionComponent && pathParts.length) {
+                    conditionComponent = flattenedComponents[`${pathParts.join('.')}`];
+                    if (!conditionComponent) {
+                        valuePathParts.unshift(pathParts.pop());
+                    }
+                }
+                if (conditionComponent && conditionComponent.type === 'selectboxes' && valuePathParts.length) {
+                    console.warn('Condition based on selectboxes has wrong format. Resave the form in the form builder to fix it.');
+                    conditionComponentPath = pathParts.join('.');
+                    comparedValue = valuePathParts.join('.');
+                }
+            }
+            // const conditionComponent = getComponent(form?.components || components, conditionComponentPath, true);
 
             const splittedConditionPath = conditionComponentPath.split('.');
             // @ts-ignore