+++ src/utils/logic.ts
@@ -3,6 +3,7 @@ import { checkCustomConditional, checkJsonConditional, checkLegacyConditional, c
 import { LogicActionCustomAction, LogicActionMergeComponentSchema, LogicActionProperty, LogicActionPropertyBoolean, LogicActionPropertyString, LogicActionValue } from "types/AdvancedLogic";
 import { get, set, clone, isEqual, assign } from 'lodash';
 import { evaluate, interpolate } from 'modules/jsonlogic';
+import { registerEphemeralState } from "./utils";
 
 export const hasLogic = (context: LogicContext): boolean => {
     const { component } = context;
@@ -41,7 +42,7 @@ export const checkTrigger = (context: LogicContext, trigger: any): boolean => {
 };
 
 export function setActionBooleanProperty(context: LogicContext, action: LogicActionPropertyBoolean): boolean {
-    const { component, scope, path, row } = context;
+    const { component, scope, path } = context;
     const property = action.property.value;
     const currentValue = get(component, property, false).toString();
     const newValue = action.state.toString();
@@ -57,11 +58,12 @@ export function setActionBooleanProperty(context: LogicContext, action: LogicAct
             if (!(scope as ConditionsScope).conditionals) {
                 (scope as ConditionsScope).conditionals = [];
             }
-            const conditionalyHidden = (scope as ConditionsScope).conditionals?.find((cond: any) => {
+            const conditionallyHidden = (scope as ConditionsScope).conditionals?.find((cond: any) => {
                 return cond.path === path
             });
-            if (conditionalyHidden) {
-                conditionalyHidden.conditionallyHidden = !!component.hidden;
+            if (conditionallyHidden) {
+                conditionallyHidden.conditionallyHidden = !!component.hidden;
+                registerEphemeralState(component, 'conditionallyHidden', !!component.hidden);
             }
             else {
                 (scope as ConditionsScope).conditionals?.push({